+++ src/simtools/testing/sim_telarray_metadata.py
@@ -6,8 +6,8 @@ import numpy as np
 
 from simtools.simtel.simtel_config_reader import SimtelConfigReader
 from simtools.simtel.simtel_config_writer import sim_telarray_random_seeds
+from simtools.simtel.simtel_io_file_info import get_corsika_run_number
 from simtools.simtel.simtel_io_metadata import (
-    get_corsika_run_number,
     get_sim_telarray_telescope_id,
     read_sim_telarray_metadata,
 )
